requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
dnslib==0.9.19
idna==3.3
numpy==1.22.4
pandas==1.4.2
PySocks==1.7.1
python-dateutil==2.8.2
python-socks==2.0.3
pytz==2022.1
requests==2.28.0
six==1.16.0
urllib3==1.26.9